+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - remembering state of Updated Files modal between opening/closing
 - remembering state of Files sections in the sidebar when switching between them
 - correctly using the worktree directory in terminal in subtasks
+- using merge-base for diff calculations in worktree mode to show only changed files since rebase
 
 ## [0.59.0]